Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/module-expected.txt

CONSOLE MESSAGE: line 1: TypeError: 'application/json' is not a valid JavaScript MIME type.

FAIL JSON modules assert_unreached: [object Event] Reached unreachable code
